+++ gym/customers/SalaryManager.java
@@ -1,7 +1,7 @@
 package gym.customers;
 
     public class SalaryManager {
-        public double calculateSalary(Instructor instructor, int numOfSessions) {
+        public int calculateSalary(Instructor instructor, int numOfSessions) {
             return (instructor.getSalary() * numOfSessions) ;
         }
     }
